requirements.txt
mysql-connector-python==9.2.0
rapidfuzz==3.11.0
requests==2.32.3
python-dotenv==1.0.1
